file reorganization
added a pages folder and resources folder
