+++ frontend/src/pages/Landing.jsx
@@ -1,7 +1,7 @@
 import { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import { useTheme } from './resources/ThemeContext';
-import Header from './Header';
+import { useTheme } from '../resources/ThemeContext';
+import Header from '../resources/Header';
 
 export default function Landing() {
   const navigate = useNavigate();
